requirements.txt
beautifulsoup4==4.7.1
panflute==1.11.2
pypandoc==1.4
python-docx==0.8.10
python-slugify==3.0.2
requests==2.21.0
